Bitmap and mman WIP - glitches at high addr

+++ Kernel/include/bitmap.h
@@ -6,12 +6,11 @@
 #include <stdio.h>
 #include <string.h>
 #include <stdlib.h>
-
 #define BITS_PER_SLOT 8
 
 typedef uint8_t bit_index;
 typedef uint16_t slot_index;
-typedef int32_t bit_amount;
+typedef int64_t bit_amount;
 
 bool bmp_initialize(uint8_t * start_address, bit_amount size_in_bits);
 void bmp_set_on(bit_amount bits_to_set, bit_amount start_bit);
